IndexedDB: Remove m_contextStopped from ActiveDOMObjecs
m_contextStopped is equivalent to !getExecutionContext().
So this CL removes m_contextStopped. The only gotcha is
that the context is cleared in the base class *after*
contextDestroyed() is called, so the one place that checked
it synchronously during the call - IDBTransaction -
needed special attention.

BUG=589507

Review-Url: https://codereview.chromium.org/2473203002
Cr-Original-Commit-Position: refs/heads/master@{#429952}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e5f1e46d7079517f3845b2ce1205c948284631b9