Fix UAF in closures posted from InProcessWorkerObjectProxy
Tasks posted by InProcessWorkerObjectProxy can be arbitrarily delayed
so it's not safe to post a closure with an crossThreadUnretained pointer
to the messaging proxy because we can't guarantee it's still there when
the closure eventually gets run.

We fix this with a WeakPtr.

BUG=660427

Review-Url: https://codereview.chromium.org/2478113002
Cr-Original-Commit-Position: refs/heads/master@{#430251}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31ed1877bdb9c27eae9d9da2a3954c16d5758a9b