Do not call internals.settings.setOverlayScrollbarsEnabled() in onload
internals.settings.setOverlayScrollbarsEnabled() doesn't work well when it is
called after ScrollbarTheme is initialized in the first layout.

The CL [1] modifies the timing of document load event and makes onload() in
fast/scrolling/scrollbar-tickmarks-hittest.html to be called after the first
layout, and thus breaks the test.

This CL moves setOverlayScrollbarsEnabled() to outside onload() such that it
is called before ScrollbarTheme is initialized.

[1] https://codereview.chromium.org/2269043002/

BUG=624697
TEST=fast/scrolling/scrollbar-tickmarks-hittest.html

Review-Url: https://codereview.chromium.org/2484613003
Cr-Original-Commit-Position: refs/heads/master@{#430284}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6253bb13131f4afaba8140133b1b865487b82006