Fix backspace double occuring on PDF files.
For some reason I overlooked an ifdef I left in some code in change;
https://chromium.googlesource.com/chromium/src/+/98ec378a9fcf5d0ace4f82d4c2e9aaafc4be492d%5E%21/

Make sure this code isn't ifdef'd out since this is the easiest path to
merge into M55. I've added a TODO to revisit this since I think this
function can actually be removed and not generate the events in the first
place.

BUG=663002

Review-Url: https://codereview.chromium.org/2481943002
Cr-Original-Commit-Position: refs/heads/master@{#430478}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 435a59e7bbcf2e5875f860317f1056d914a30610