Remove InterpolableBool and existing use of it in SVG path interpolation
The InterpolableBool class should not be used as bools are not considered
interpolable. This change replaces the single use of it with
InterpolableNumber. This retains the existing behaviour but is not
ideal.

This change fixes a nullptr crash caused by the use of InterpolableBool.

BUG=662694

Review-Url: https://codereview.chromium.org/2487913002
Cr-Original-Commit-Position: refs/heads/master@{#430907}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2439d975210aba96119d8b75633b9f5659563d00