Clean up after the Attr odyssey
Attr once again inherits from Node in the spec:
https://github.com/whatwg/dom/pull/299

Remove use counters used to gague the risk of the spec as it was, and
remove the nodeValue and textContent overrides used to support them.
That in turn required some internal changes to preserve behavior.

BUG=305105

Review-Url: https://codereview.chromium.org/2365563002
Cr-Original-Commit-Position: refs/heads/master@{#430909}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 08e70eafd07278df1ff50d8c26055c884e13825e