Part 2.2: Is policy list subsumed under subsuming policy?
This is part of an experimental feature Embedding-CSP.

In particular this is the second CL devoted to implementing
*3.3. Is policy list subsumed under subsuming policy?

Here we try to find whether a source list directive is subsuming
to a vector of source list directives.

Note: right now we only consider non-special keyword cases, i.e.
CSPSources.

BUG=647588

Review-Url: https://codereview.chromium.org/2452903004
Cr-Original-Commit-Position: refs/heads/master@{#431549}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7650479aed19de17e02478abd1845ffea51e7dbf