Changing spell check related IDL to add RaiseException attribute
Add RaiseExpcetion to Internals.hasSpellingMarker(), hasGrammerMarker(), and
setSpellCheckingEnabled().

This patch is to improve the spell check related methods in Internals. We change
ASSERT to RaiseExpcetion since ASSERT is not allowed to use in presubmit. And
yosin@ suggestion us use RaiseExpcetion replace these. Discussion is in
https://codereview.chromium.org/2457523003/

BUG=none

Review-Url: https://codereview.chromium.org/2498493002
Cr-Original-Commit-Position: refs/heads/master@{#431569}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d79a033a04f9b8994472b871254d17269b8464eb