Remove gvr SDK dependency
SDK is used to only get a string which only avaialble in SDK. A bug has filed
to make the string available in NDK too. Before it fixed, we use a hard coded
string to remove the SDK dependency. This should reduce the binary size since
the proguard is configured to keep a lot of java functions.

BUG=664145

Review-Url: https://codereview.chromium.org/2490003003
Cr-Original-Commit-Position: refs/heads/master@{#431633}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c84c42ec3910d27eea80a9bef3db419339aff884