DevToosl: bring securityIcons.svg up-to-date with actual png files.
The securityIcons.png we use today is *not* generated out of the SVG file.
The source was provided separately from the PNG files; the PNG files
were *not* generated with our toolchain.

This patch brings securityIcons.svg up to date and feeds it to our image
toolchain to re-generate PNG files.

BUG=none
R=dgozman

Review-Url: https://codereview.chromium.org/2485853002
Cr-Original-Commit-Position: refs/heads/master@{#431694}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3bb3a1997a2b51f504676d76f0148974736e4630