binding: Makes Dictionary handle a possible exception in [[Get]].
Dictionary retrieves a dictionary member through the [[Get]] internal
method of the dictionary object.  Since the properties can be accessor
properties, arbitrary script may run in [[Get]], so we have to handle
a possible exception in [[Get]].

The right fix is to propagate the exception up to the caller,
however, there are so many callers and the issue is urgent.
So simply swallowing all the exceptions in this CL, and fix it
re-throws in a follow-up CL.

BUG=644237

Review-Url: https://codereview.chromium.org/2496533002
Cr-Original-Commit-Position: refs/heads/master@{#431848}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f971b7bafe1c9a8574e66525b593644e4b9d658