binding: Replaces DictionaryHelper::get(ScriptValue) with get(Nullable<T>).
Replaces DictionaryHelper::get(dict, key, ScriptValue&) with
get(dict, key, Nullable<T>&).

The only client of get(ScriptValue&) is EffectInput.cpp and it
expects "double?" in IDL.  So, Nullable<T> should be better.

The motivation of this CL is to remove an use of
ScriptState::current() in order to make Blink use the correct
context.  Also we'd like to reduce usage of ScriptValue that
saves a ScriptState inside.

BUG=618672

Review-Url: https://codereview.chromium.org/2444363002
Cr-Original-Commit-Position: refs/heads/master@{#431876}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 05d7c9d9594f75024b20a53cfbb1d6a8f5deb938