[draft] PlzNavigate: send searchable form data to the browser side on form submission.
In the non-PlzNavigate case, searchable form data is stored in the provisional
document loader by RenderFrameImpl::willSubmitForm, and later sent to the
browser in the DidCommitProvisionalLoad message.

In the PlzNavigate case, however, there isn't a provisional document loader when
RenderFrameImpl::willSubmitForm is called. Therefore, this CL sends the
information to the browser side in the BeginNavigation message, and store it in
NavigationHandleImpl.

This fixes TemplateURLScraperTest.ScrapeWithOnSubmit for PlzNavigate.

BUG=504347
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2485783002
Cr-Original-Commit-Position: refs/heads/master@{#431915}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15ceb3909b88d7d858a47547c9fc03f1eee20776