Make UA composition events composed events
This is a bug fix, as we did it for DOMActivate event in
https://codereview.chromium.org/2491233002.

It looks that we forgot to make composition events
(compositionstart/compositionupdate/compositionend) composed events.

This CL also add a test for all UI events, including legacy UI Events, which should
be composed events.

Review-Url: https://codereview.chromium.org/2498263002
Cr-Original-Commit-Position: refs/heads/master@{#432127}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f508ad541bff707485fcfba84a1c20379f0f7589